+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/transaction/impl/DefaultCommitTimestampLoaderTest.java
@@ -45,7 +45,7 @@ import org.eclipse.collections.api.factory.primitive.LongLists;
 import org.eclipse.collections.api.map.primitive.LongLongMap;
 import org.junit.jupiter.api.Test;
 
-public class CommitTimestampLoaderTest {
+public class DefaultCommitTimestampLoaderTest {
     private static final TableReference TABLE_REF = TableReference.fromString("table");
     private final TimestampCache timestampCache = mock(TimestampCache.class);
     private final TransactionConfig transactionConfig = mock(TransactionConfig.class);
@@ -77,7 +77,8 @@ public class CommitTimestampLoaderTest {
         setup(startTs, commitTs);
 
         // no immutableTs lock for read-only transaction
-        CommitTimestampLoader commitTimestampLoader = getCommitTsLoader(Optional.empty(), transactionTs, commitTs - 1);
+        DefaultCommitTimestampLoader commitTimestampLoader =
+                getCommitTsLoader(Optional.empty(), transactionTs, commitTs - 1);
 
         assertCanGetCommitTs(startTs, commitTs, commitTimestampLoader);
     }
@@ -91,7 +92,7 @@ public class CommitTimestampLoaderTest {
         setup(startTs, commitStatus, true);
 
         // no immutableTs lock for read-only transaction
-        CommitTimestampLoader commitTimestampLoader =
+        DefaultCommitTimestampLoader commitTimestampLoader =
                 getCommitTsLoader(Optional.empty(), transactionTs, transactionTs + 1);
 
         assertThatExceptionOfType(ExecutionException.class)
@@ -111,7 +112,7 @@ public class CommitTimestampLoaderTest {
         setup(startTs, commitTs);
 
         // no immutableTs lock for read-only transaction
-        CommitTimestampLoader commitTimestampLoader =
+        DefaultCommitTimestampLoader commitTimestampLoader =
                 getCommitTsLoader(Optional.empty(), transactionTs, transactionTs + 1);
 
         assertCanGetCommitTs(startTs, commitTs, commitTimestampLoader);
@@ -128,7 +129,7 @@ public class CommitTimestampLoaderTest {
         LockToken lock = mock(LockToken.class);
 
         // no immutableTs lock for read-only transaction
-        CommitTimestampLoader commitTimestampLoader =
+        DefaultCommitTimestampLoader commitTimestampLoader =
                 getCommitTsLoader(Optional.of(lock), transactionTs, transactionTs + 1);
 
         // the transaction will eventually throw at commit time. In this test we are only concerned with per read
@@ -147,7 +148,8 @@ public class CommitTimestampLoaderTest {
         LockToken lock = mock(LockToken.class);
 
         // no immutableTs lock for read-only transaction
-        CommitTimestampLoader commitTimestampLoader = getCommitTsLoader(Optional.of(lock), transactionTs, commitTs + 1);
+        DefaultCommitTimestampLoader commitTimestampLoader =
+                getCommitTsLoader(Optional.of(lock), transactionTs, commitTs + 1);
         assertCanGetCommitTs(startTs, commitTs, commitTimestampLoader);
     }
 
@@ -161,7 +163,7 @@ public class CommitTimestampLoaderTest {
         long startTsUnknown = 7l;
         TransactionStatus commitUnknown = TransactionStatus.unknown();
 
-        CommitTimestampLoader commitTimestampLoader =
+        DefaultCommitTimestampLoader commitTimestampLoader =
                 getCommitTsLoader(Optional.empty(), transactionTs, transactionTs - 1);
 
         setup(startTsKnown, commitTsKnown);
@@ -180,7 +182,7 @@ public class CommitTimestampLoaderTest {
         verifyNoMoreInteractions(timestampCache);
     }
 
-    private void assertCanGetCommitTs(long startTs, long commitTs, CommitTimestampLoader commitTimestampLoader)
+    private void assertCanGetCommitTs(long startTs, long commitTs, DefaultCommitTimestampLoader commitTimestampLoader)
             throws InterruptedException, ExecutionException {
         LongLongMap loadedCommitTs = commitTimestampLoader
                 .getCommitTimestamps(TABLE_REF, LongLists.immutable.of(startTs), false, transactionService)
@@ -189,10 +191,10 @@ public class CommitTimestampLoaderTest {
         assertThat(loadedCommitTs.get(startTs)).isEqualTo(commitTs);
     }
 
-    private CommitTimestampLoader getCommitTsLoader(
+    private DefaultCommitTimestampLoader getCommitTsLoader(
             Optional<LockToken> lock, long transactionTs, long lastSeenCommitTs) {
         createKnowledgeComponents(lastSeenCommitTs);
-        CommitTimestampLoader commitTimestampLoader = new CommitTimestampLoader(
+        DefaultCommitTimestampLoader commitTimestampLoader = new DefaultCommitTimestampLoader(
                 timestampCache,
                 lock, // commitTsLoader does not care if the lock expires.
                 () -> transactionTs,
