+++ atlasdb-tests-shared/src/test/java/com/palantir/atlasdb/transaction/impl/SnapshotTransactionManagerTest.java
@@ -108,7 +108,7 @@ public class SnapshotTransactionManagerTest {
                 mock(TransactionService.class),
                 () -> AtlasDbConstraintCheckingMode.FULL_CONSTRAINT_CHECKING_THROWS_EXCEPTIONS,
                 null,
-                null,
+                SweepStrategyManagers.createDefault(keyValueService),
                 cleaner,
                 false,
                 TransactionTestConstants.GET_RANGES_THREAD_POOL_SIZE,
@@ -309,7 +309,7 @@ public class SnapshotTransactionManagerTest {
                 mock(TransactionService.class),
                 () -> null,
                 null,
-                null,
+                SweepStrategyManagers.createDefault(keyValueService),
                 cleaner,
                 false,
                 TransactionTestConstants.GET_RANGES_THREAD_POOL_SIZE,
